🧹 move `inmemory/kiss.go` -> `inmemory/store/kiss.go`
